requirements.txt
ruamel.yaml; python_version >= '3.6'
paramiko; python_version >= '3.6'
decorator; python_version >= '3.6'
